implementacao-tecnica/backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==3.2.2
certifi==2025.11.12
cffi==2.0.0
click==8.1.8
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.115.0
gotrue==2.12.4
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
hyperframe==6.1.0
idna==3.11
jiter==0.12.0
openai==2.8.1
packaging==25.0
passlib==1.7.4
postgrest==0.16.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.1
pypdf==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.9
PyYAML==6.0.3
realtime==1.0.6
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.38.6
storage3==0.7.7
StrEnum==0.4.15
supabase==2.6.0
supafunc==0.5.1
tqdm==4.67.1
typing_extensions==4.15.0
uvicorn==0.30.6
uvloop==0.22.1
watchfiles==1.1.1
websockets==12.0
